NavigationQRShareButton refactored
