+++ VultisigApp/VultisigApp/Views/Components/Navigation Items/NavigationQRShareButton.swift
@@ -24,17 +24,7 @@ struct NavigationQRShareButton: View {
     @State var imageName: String = ""
     
     var body: some View {
-#if os(iOS)
-        ZStack {
-            if type == .Address {
-                shareLink
-            } else {
-                button
-            }
-        }
-#elseif os(macOS)
-        shareLink
-#endif
+        container
     }
     
     var shareLink: some View {
@@ -46,16 +36,6 @@ struct NavigationQRShareButton: View {
             }
         }
     }
-    
-#if os(iOS)
-    var button: some View {
-        Button {
-            shareImage()
-        } label: {
-            content
-        }
-    }
-#endif
     
     private func getLink(image: Image) -> some View {
         ShareLink(
@@ -95,48 +75,6 @@ struct NavigationQRShareButton: View {
             imageName = "\(suffix)-\(vault.name)-\(id.suffix(3))-\(date).png"
         }
     }
-    
-#if os(iOS)
-    func shareImage() {
-        guard let image = renderedImage else {
-            return
-        }
-        
-        let uiImage = renderImage(image: image)
-
-        let tempURL = FileManager.default.temporaryDirectory.appendingPathComponent(imageName)
-
-        if let pngData = uiImage.pngData() {
-            try? pngData.write(to: tempURL)
-        }
-
-        let activityViewController = UIActivityViewController(activityItems: [tempURL], applicationActivities: nil)
-
-        if let windowScene = UIApplication.shared.connectedScenes.first as? UIWindowScene,
-           let rootViewController = windowScene.windows.first?.rootViewController {
-            rootViewController.present(activityViewController, animated: true, completion: nil)
-        }
-    }
-    
-    private func renderImage(image: Image) -> UIImage {
-        let controller = UIHostingController(
-            rootView:
-                image
-                    .frame(width: 278, height: 500)
-                    .offset(y: -30)
-        )
-        let view = controller.view
-
-        let targetSize = controller.view.intrinsicContentSize
-        view?.bounds = CGRect(origin: .zero, size: targetSize)
-        view?.backgroundColor = .clear
-
-        let renderer = UIGraphicsImageRenderer(size: targetSize)
-        return renderer.image { _ in
-            view?.drawHierarchy(in: controller.view.bounds, afterScreenUpdates: true)
-        }
-    }
-#endif
 }
 
 #Preview {
